Final changes sprint 1

+++ app.py
@@ -153,15 +153,28 @@ def search():
 
 @app.route('/addproduct', methods=['POST'])
 def add_product():
-    vendor_id = request.json["vendor_id"]
+    if not ("name" in request.json and
+            "description" in request.json and
+            "price" in request.json and
+            "stock" in request.json and
+            "category" in request.json and
+            "image" in request.json):
+        abort(400)
+    token = extract_auth_token(request)
+    try:
+        vendor_id = decode_token(token)
+    except (jwt.ExpiredSignatureError, jwt.InvalidTokenError):
+        abort(403)
+    #vendor_id = request.json["vendor_id"]
+    name = request.json["name"]
     description = request.json["description"]
     price = request.json["price"]
     stock = request.json["stock"]
     category = request.json["category"]
     image = request.json["image"]
 
-    if not (vendor_id and description and price and stock and category and image):
-        return
+    # if not (vendor_id and name and description and price and stock and category and image):
+    #     abort(400)
     
     # Check if profile is a vendor
     v = User.query.filter_by(user_id=vendor_id).first()
@@ -169,7 +182,7 @@ def add_product():
     if not v or v.role != UserRole.VENDOR:
         abort(403)
     
-    p = Product(vendor_id, v.username, description, price, stock, category, image, 0)
+    p = Product(vendor_id, name, description, price, stock, category, image, 0)
 
     db.session.add(p)
     db.session.commit()
@@ -179,10 +192,17 @@ def add_product():
 
 @app.route('/deleteproduct', methods=['POST'])
 def delete_product():
-    product_id = request.json["product_id"]
-
-    if not product_id:
+    token = extract_auth_token(request)
+    try:
+        vendor_id = decode_token(token)
+    except (jwt.ExpiredSignatureError, jwt.InvalidTokenError):
         abort(403)
+    product_id = request.json["product_id"]
+    v = User.query.filter_by(user_id=vendor_id).first()
+    if not v or v.role != UserRole.VENDOR:
+        abort(403)
+    if not product_id:
+        abort(400)
         
     Review.query.filter_by(product_id=product_id).delete()
     Product.query.filter_by(product_id=product_id).delete()
@@ -193,43 +213,49 @@ def delete_product():
 
 @app.route('/changeproduct', methods=['POST'])
 def change_product():
+    token = extract_auth_token(request)
+    try:
+        vendor_id = decode_token(token)
+    except (jwt.ExpiredSignatureError, jwt.InvalidTokenError):
+        abort(403)
     product_id = request.json["product_id"]
     description = request.json["description"]
     price = request.json["price"]
     stock = request.json["stock"]
-    category = request.json["category"]
+    #category = request.json["category"]
     image = request.json["image"]
-    rating = request.json["rating"]
-
-    if not product_id:
+    #rating = request.json["rating"]
+    vendor = User.query.filter_by(user_id=vendor_id).first()
+    if not vendor or vendor.role != UserRole.VENDOR:
         abort(403)
+    if not product_id:
+        abort(400)
 
     p = Product.query.filter_by(product_id=product_id).first()
 
     if not p:
-        abort(403)
+        abort(400)
 
     # Change information
     if description: p.description = description
     if price: p.price = price
     if stock: p.stock = stock
-    if category: p.category = category
-    if image: p.image = image
+    if image: p.image_url = image
 
     # Update Rating by calculating all reviews with that product id
-    r = Review.query.filter_by(product_id=product_id).all()
+    # r = Review.query.filter_by(product_id=product_id).all()
 
-    if rating>=0 and len(r) > 0:
-        n = len(r) + 1          # To account for the new rating as well
-        s = 0
-        for i in r:
-            s += i.rating/n
-        s+=rating/n
+    # if rating>=0 and len(r) > 0:
+    #     n = len(r) + 1          # To account for the new rating as well
+    #     s = 0
+    #     for i in r:
+    #         s += i.rating/n
+    #     s+=rating/n
 
-        p.rating = s
+    #     p.rating = s
 
-    elif rating>=0 and len(r) == 0:
-        p.rating = rating
+    # elif rating>=0 and len(r) == 0:
+    #     p.rating = rating
 
     db.session.commit()
